models/requirements.txt
fastapi
uvicorn[standard]
pillow
ultralytics
numpy
opencv-python
roboflow
tqdm